refactor: fix naming convention
