+++ lib/main.dart
@@ -1,15 +1,9 @@
-import 'package:fiservonboardingexp/config/app_router.dart';
 import 'package:fiservonboardingexp/screens/Login_page.dart';
-import 'package:fiservonboardingexp/screens/home_page.dart';
-import 'package:fiservonboardingexp/screens/task_categories/compliance_page.dart';
-import 'package:fiservonboardingexp/screens/task_categories/week_one_page.dart';
-import 'package:fiservonboardingexp/screens/training_page.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:fiservonboardingexp/firebase_options.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:fiservonboardingexp/screens/teaser.dart';
-import 'package:fiservonboardingexp/screens/nav_bar.dart';
 import 'screens/main_screen.dart';
 
 int? initScreen;
@@ -34,7 +28,7 @@ class MyApp extends StatelessWidget {
       ),
       initialRoute: initScreen == 0 || initScreen == null ? 'teaser' : 'home',
       routes: {
-        'home': (context) => MainScreen(),
+        'home': (context) => const LoginApp(),
         'teaser': (context) => const TeaserScreen(),
       },
     );
