+++ lib/screens/faq_page_placeholder.dart
@@ -1,4 +1,4 @@
-import 'package:fiservonboardingexp/screens/nav_bar.dart';
+import 'package:fiservonboardingexp/widgets/nav_bar.dart';
 import 'package:flutter/material.dart';
 import 'app_bar.dart';
 
@@ -7,7 +7,7 @@ class FaqPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return const Scaffold(
       appBar: AppBarOverlay(),
-      bottomNavigationBar: customNavBar(),
+      bottomNavigationBar: CustomNavBar(),
       body: Padding(
         padding: EdgeInsets.only(top: 16, left: 16),
         child: Align(
