+++ lib/screens/home_page.dart
@@ -1,4 +1,4 @@
-import 'package:fiservonboardingexp/screens/nav_bar.dart';
+import 'package:fiservonboardingexp/widgets/nav_bar.dart';
 
 import 'app_bar.dart';
 import 'package:flutter/material.dart';
@@ -13,7 +13,7 @@ class HomePage extends StatelessWidget {
       home: Scaffold(
         // //appbar
         appBar: const AppBarOverlay(),
-        bottomNavigationBar: customNavBar(),
+        bottomNavigationBar: const CustomNavBar(),
         //Home Page
         body: Scrollbar(
             thumbVisibility: true,
