+++ lib/screens/modules_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'nav_bar.dart';
+import '../widgets/nav_bar.dart';
 
 class ModulesPage extends StatelessWidget {
   const ModulesPage({super.key});
@@ -8,7 +8,7 @@ class ModulesPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return const Scaffold(
-      bottomNavigationBar: const customNavBar(),
+      bottomNavigationBar: CustomNavBar(),
       body: Padding(
         padding: EdgeInsets.only(top: 16, left: 16),
         child: Align(
