+++ lib/screens/task_categories/health_and_safety_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import '../../util/task_tile.dart';
 import '../../util/constants.dart';
-import '../nav_bar.dart';
+import '../../widgets/nav_bar.dart';
 
 class HnSTasks extends StatefulWidget {
   const HnSTasks({Key? key}) : super(key: key);
@@ -18,7 +18,7 @@ class _HnSPage extends State<HnSTasks> {
       debugShowCheckedModeBanner: false,
       home: Scaffold(
         appBar: myAppBar,
-        bottomNavigationBar: const customNavBar(),
+        bottomNavigationBar: const CustomNavBar(),
         body: Container(
           color: Colors.white,
           child: Padding(
