+++ lib/util/constants.dart
@@ -1,4 +1,4 @@
-import 'package:fiservonboardingexp/screens/nav_bar.dart';
+import 'package:fiservonboardingexp/widgets/nav_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:google_nav_bar/google_nav_bar.dart';
 import '../screens/app_bar.dart';
@@ -28,7 +28,7 @@ const myDrawer = Drawer(
   ]),
 );
 
-final navBar = customNavBar;
+final navBar = CustomNavBar();
 
 const myBackgroundColor = Color.fromARGB(255, 39, 39, 39);
 const fiservColor = Color.fromARGB(255, 255, 102, 0);
